fix: add dangledFilesService to manage dangling files in backups
- Introduced DangledFilesService to handle the detection and processing of dangling files.
- Refactored Backup class to utilize DangledFilesService for resyncing files.
- Removed outdated file downloadability logic from Backup class.
- Updated dependency injection to register DangledFilesService.
