+++ src/apps/backups/dependency-injection/BackupsDependencyContainerFactory.ts
@@ -8,6 +8,7 @@ import { Backup } from '../Backups';
 import { registerLocalTreeServices } from './local/registerLocalTreeServices';
 import { registerRemoteTreeServices } from './virtual-drive/registerRemoteTreeServices';
 import { registerUserUsageServices } from './user/registerUsageServices';
+import { DangledFilesService } from '../dangled-files/DangledFilesService';
 
 @Service()
 export class BackupsDependencyContainerFactory {
@@ -35,6 +36,9 @@ export class BackupsDependencyContainerFactory {
 
       Logger.info('[BackupsDependencyContainerFactory] Registering user usage services.');
       await registerUserUsageServices(builder);
+
+      Logger.info('[BackupsDependencyContainerFactory] Registering dangled files service.');
+      builder.registerAndUse(DangledFilesService);
 
       Logger.info('[BackupsDependencyContainerFactory] Registering Backup service.');
       await builder.registerAndUse(Backup);
